refactor: 쿠키 관련 서비스는 AuthCookieService로 분리
