+++ src/main/java/com/keeper/homepage/global/config/security/filter/RefreshTokenFilter.java
@@ -4,8 +4,8 @@ import static com.keeper.homepage.global.config.security.data.JwtType.ACCESS_TOK
 import static com.keeper.homepage.global.config.security.data.JwtType.REFRESH_TOKEN;
 import static com.keeper.homepage.global.config.security.data.JwtValidationType.EXPIRED;
 
+import com.keeper.homepage.domain.auth.application.AuthCookieService;
 import com.keeper.homepage.global.config.security.JwtTokenProvider;
-import com.keeper.homepage.global.config.security.data.JwtType;
 import com.keeper.homepage.global.config.security.data.JwtValidationType;
 import com.keeper.homepage.global.config.security.data.TokenValidationResultDto;
 import com.keeper.homepage.global.util.redis.RedisUtil;
@@ -13,7 +13,6 @@ import jakarta.servlet.FilterChain;
 import jakarta.servlet.ServletException;
 import jakarta.servlet.ServletRequest;
 import jakarta.servlet.ServletResponse;
-import jakarta.servlet.http.Cookie;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import java.io.IOException;
@@ -32,6 +31,7 @@ public class RefreshTokenFilter extends GenericFilterBean {
 
   private final JwtTokenProvider jwtTokenProvider;
   private final RedisUtil redisUtil;
+  private final AuthCookieService authCookieService;
 
   @Override
   public void doFilter(ServletRequest request, ServletResponse response, FilterChain filterChain)
@@ -55,11 +55,7 @@ public class RefreshTokenFilter extends GenericFilterBean {
       String[] roles = jwtTokenProvider.getRoles(refreshTokenDto.getToken());
       HttpServletResponse httpResponse = (HttpServletResponse) response;
 
-      String newRefreshToken = jwtTokenProvider.createAccessToken(REFRESH_TOKEN, authId, roles);
-      setTokenInCookie(REFRESH_TOKEN, httpResponse, newRefreshToken);
-      String newAccessToken = jwtTokenProvider.createAccessToken(ACCESS_TOKEN, authId, roles);
-      setTokenInCookie(ACCESS_TOKEN, httpResponse, newAccessToken);
-      redisUtil.setDataExpire(authId, newRefreshToken, REFRESH_TOKEN.getExpiredMillis());
+      authCookieService.setNewCookieInResponse(authId, roles, httpResponse);
     }
   }
 
@@ -75,12 +71,5 @@ public class RefreshTokenFilter extends GenericFilterBean {
 
   private boolean isRefreshTokenValid(TokenValidationResultDto refreshTokenDto) {
     return refreshTokenDto.isValid();
-  }
-
-  private void setTokenInCookie(JwtType jwtType, HttpServletResponse httpResponse, String token) {
-    Cookie cookie = new Cookie(jwtType.getTokenName(), token);
-    cookie.setHttpOnly(true);
-    cookie.setMaxAge((int) jwtType.getExpiredMillis() / 1000);
-    httpResponse.addCookie(cookie);
   }
 }
